requirements.txt
django
django-rest-framework
djoser
